__WIP__ because I'm still tracking down some performance nits and working on tests. This uses the newly-added SSE2 and BMI1 intrinsics to optimize the implementation of `UTF8Encoding.GetCharCount` (which is also used as the internal workhorse for `Utf8String.IsWellFormed`).

For texts which are primarily a mixture of ASCII and 2-byte subsequences (Greek, Cyrillic, Western European languages with accented characters, etc.), this provides a marked performance improvement in the validation APIs. See in particular the last two rows of the below table.

There's some noise while running the tests. I'm not quite sure what's causing this, but it seems my earlier concerns about this regressing 3-byte processing may not have been warranted.

|        Method | Toolchain |      Corpus |       Mean |     Error |    StdDev | Ratio |
|-------------- |---------- |------------ |-----------:|----------:|----------:|------:|
| **Utf8_Is_Valid** |    **master** |    **11-0.txt** |  **42.041 us** | **0.3183 us** | **0.2977 us** |  **1.00** |
| Utf8_Is_Valid |     proto |    11-0.txt |  42.612 us | 0.4118 us | 0.3852 us |  1.01 |
| Utf8_Is_Valid |    proto2 |    11-0.txt |  40.538 us | 0.5063 us | 0.4228 us |  0.96 |
|               |           |             |            |           |           |       |
| **Utf8_Is_Valid** |    **master** |      **11.txt** |   **3.462 us** | **0.0172 us** | **0.0144 us** |  **1.00** |
| Utf8_Is_Valid |     proto |      11.txt |   3.376 us | 0.0440 us | 0.0390 us |  0.98 |
| Utf8_Is_Valid |    proto2 |      11.txt |   3.412 us | 0.0152 us | 0.0143 us |  0.98 |
|               |           |             |            |           |           |       |
| **Utf8_Is_Valid** |    **master** | **25249-0.txt** |  **72.546 us** | **0.3310 us** | **0.2764 us** |  **1.00** |
| Utf8_Is_Valid |     proto | 25249-0.txt |  66.819 us | 0.3117 us | 0.2434 us |  0.92 |
| Utf8_Is_Valid |    proto2 | 25249-0.txt |  69.873 us | 0.3105 us | 0.2593 us |  0.96 |
|               |           |             |            |           |           |       |
| **Utf8_Is_Valid** |    **master** | **30774-0.txt** |  **76.668 us** | **0.9464 us** | **0.8390 us** |  **1.00** |
| Utf8_Is_Valid |     proto | 30774-0.txt |  36.461 us | 0.3560 us | 0.3330 us |  0.48 |
| Utf8_Is_Valid |    proto2 | 30774-0.txt |  30.343 us | 0.5120 us | 0.4276 us |  0.40 |
|               |           |             |            |           |           |       |
| **Utf8_Is_Valid** |    **master** | **39251-0.txt** | **133.060 us** | **2.5572 us** | **2.3920 us** |  **1.00** |
| Utf8_Is_Valid |     proto | 39251-0.txt |  51.215 us | 1.0158 us | 0.9502 us |  0.39 |
| Utf8_Is_Valid |    proto2 | 39251-0.txt |  42.564 us | 0.4976 us | 0.4156 us |  0.32 |