@GrabYourPitchforks 2019-11-07
Still need to respond to PR feedback, but I'm on track to commit this to coreclr by Mon 11 Nov, ahead of the consolidation deadline.