@GrabYourPitchforks 2019-11-11
I'm holding off on checking this in for now because I'm still experimenting with various codegen patterns. Migrated to https://github.com/GrabYourPitchforks/coreclr/pull/3 so it doesn't get lost in the shuffle.